feat: add login, signup, and register with AWS cognito
This commit adds AWS cognito links to use the hosted UI for sign in, sign up, sso, and logout.
The header buttons and buttons on the videos are updated.
In the IdentityProvider is also a route for sso/login to handle the auth_token exchange
and store the access_token and use data from the identity_token.
There is also a route for sso/logout there.